Correction in VMR9MonitorInfo. hMon is a IntPtr.
git-svn-id: https://svn.code.sf.net/p/directshownet/code/trunk@998 b2a09c84-0b36-484d-b36e-ffab392acda3

+++ directshowlib/src/Vmr9.cs
@@ -253,7 +253,7 @@ namespace DirectShowLib
     {
         public int uDevID;
         public DsRect rcMonitor;
-        public int hMon;
+        public IntPtr hMon;
         public int dwFlags;
         [MarshalAs(UnmanagedType.ByValTStr, SizeConst=32)] public string szDevice;
         [MarshalAs(UnmanagedType.ByValTStr, SizeConst=512)] public string szDescription;
